module_D/D4/pyproject.toml
[tool.poetry]
name = "d3"
version = "0.1.0"
description = ""
authors = ["Nikita Baranov <nbaranov.vrn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
